aula_3/T3/flask_customer_api/requirements.txt
flask==2.3.3
python-dateutil==2.8.2
validate-docbr==1.10.0
Werkzeug==2.3.7